Fix types in events

+++ src/interfaces/IAuction.sol
@@ -69,7 +69,7 @@ interface IAuction is
     /// @param price The price of the bid
     /// @param exactIn Whether the bid is exact in
     /// @param amount The amount of the bid
-    event BidSubmitted(uint256 indexed id, address indexed owner, uint256 price, bool exactIn, uint256 amount);
+    event BidSubmitted(uint256 indexed id, address indexed owner, uint256 price, bool exactIn, uint128 amount);
 
     /// @notice Emitted when a new checkpoint is created
     /// @param blockNumber The block number of the checkpoint
@@ -77,7 +77,7 @@ interface IAuction is
     /// @param totalCleared The total amount of tokens cleared
     /// @param cumulativeMps The cumulative percentage of total tokens allocated across all previous steps, represented in ten-millionths of the total supply (1e7 = 100%)
     event CheckpointUpdated(
-        uint256 indexed blockNumber, uint256 clearingPrice, uint256 totalCleared, uint24 cumulativeMps
+        uint256 indexed blockNumber, uint256 clearingPrice, uint128 totalCleared, uint24 cumulativeMps
     );
 
     /// @notice Emitted when a bid is exited
@@ -89,7 +89,7 @@ interface IAuction is
     /// @notice Emitted when a bid is claimed
     /// @param owner The owner of the bid
     /// @param tokensFilled The amount of tokens claimed
-    event TokensClaimed(address indexed owner, uint256 tokensFilled);
+    event TokensClaimed(address indexed owner, uint128 tokensFilled);
 
     /// @notice Submit a new bid
     /// @param maxPrice The maximum price the bidder is willing to pay
